Add very hacky fix to iframe back button issue
When the URL of an iframe changes, an entry is pushed onto the history
stack:
  https://github.com/whatwg/html/issues/6501

We get two history entries if we change gene page.  Then we need to
oush the back button twice to go back to the previous gene page.  And
the first click shows the wrong iframe contents.

This hack recreate the iframe component on page change so it's not
added to the history.  Sort of like this: https://stackoverflow.com/a/79096750

Refs pombase/website#2284